refactor(cli): consolidate the two vz dev paths into one; drop dead apple_container names + Swift bridge
