--- crates/mvm-backend/swift/.gitignore
@@ -1,4 +1,0 @@
-
-.build/
-.swiftpm/
-Package.resolved
